Threading rebase (#6)
* Update http_downloader.py

* Update base_downloader.py

* fixed fatal bug

* added ETA

* Update http_downloader.py

should fix windows issues

* pretty print

* Update base_downloader.py

* Update http_downloader.py

* Update http_downloader.py

* Update http_downloader.py

* Update http_downloader.py

* Update http_downloader.py

* Update http_downloader.py

* Update http_downloader.py

* Update http_downloader.py

* Update http_downloader.py

* Update http_downloader.py

* Update base_downloader.py

* Update http_downloader.py

* Update http_downloader.py

+++ anime_downloader/downloader/base_downloader.py
@@ -35,19 +35,28 @@ class BaseDownloader:
         if self.source.referer:
             headers['referer'] = self.source.referer
 
-        # using session downloads the whole file, essentially freezing the program.
-        r = requests.get(
-            self.source.stream_url, headers=headers, stream=True)
+        for i in range(5):
+            with requests.get(self.source.stream_url, headers=headers, stream=True, verify=False) as r:
+                self._total_size = max(int(r.headers.get('Content-length', 0)), 
+                                        int(r.headers.get('Content-Length', 0)), 
+                                        int(r.headers.get('content-length', 0)))
+                if not self._total_size and not r.headers.get('Transfer-Encoding') == 'chunked':
+                    continue
 
-        self._total_size = int(r.headers['Content-length'])
-        logger.debug('total size: ' + str(self._total_size))
-        if os.path.exists(self.path):
-            if abs(os.stat(self.path).st_size - self._total_size) < 10 \
-               and not self.force:
-                logger.warning('File already downloaded. Skipping download.')
+                logger.debug('Total size: ' + str(self._total_size))
+                if os.path.exists(self.path):
+                    if abs(os.stat(self.path).st_size - self._total_size) < 10 \
+                       and not self.force:
+                        logger.warning('File already downloaded. Skipping download.')
+                        return True
+                    else:
+                        os.remove(self.path)
                 return
-            else:
-                os.remove(self.path)
+
+        if not self._total_size:
+            logger.error('Unable get the file.')
+            sys.exit(1)
+
 
     def download(self):
         # TODO: Clean this up
@@ -56,9 +65,11 @@ class BaseDownloader:
 
         # TODO: Use pathlib. Break into functions
         util.make_dir(self.path.rsplit('/', 1)[0])
-        self.check_if_exists()
 
-        self.start_time = time.time()
+        # Goes to the next episode if the file already exists.
+        if self.check_if_exists():
+            return
+
         self.downloaded = 0
         self._download()
         self.post_process()
@@ -83,8 +94,20 @@ def write_status(downloaded, total_size, start_time):
     downloaded = float(downloaded)/1048576
     total_size = float(total_size)/1048576
 
-    status = 'Downloaded: {0:.2f}MB/{1:.2f}MB, Rate: {2:.2f}KB/s'.format(
-        downloaded, total_size, rate)
+    eta = ((total_size-downloaded)*1024)/rate
+    minutes = '0' * (round((eta-(eta%60))/60) < 10) + str(round((eta-(eta%60))/60))
+    seconds = '0' * (round(eta%60) < 10) + str(round(eta%60))
+    eta = f'{minutes}:{seconds}'
+    if downloaded >= total_size:
+        eta = 'Done'
+
+    if total_size:
+        status = 'Downloaded: {0:.2f}MB/{1:.2f}MB, Rate: {2:.2f}KB/s, ETA: {3}'.format(
+            downloaded, total_size, rate, eta)
+
+    # Chunked transfer, unknown size?
+    else:
+        status = 'Downloaded: {0:.2f}MB, Rate: {1:.2f}KB/s'.format(downloaded, rate)
 
     sys.stdout.write("\r" + status + " "*5 + "\r")
     sys.stdout.flush()
